Revert operator addresses to mainnet
I have reverted every unknown operator address to the mainnet address if there was an obvious mainnet address.

The reasoning is as follows.

An important goal for this testnet is to understand if custodial infrastructure is working post startgate. If we swap operator addresses, we lose the opportunity to learn about this problems because uses will happily use the gaia keyring rather than secure storage.